Moved getConversationDetails to new file
