+++ front/lib/notifications/workflows/conversation-unread.ts
@@ -9,10 +9,6 @@ import {
 import { getSmallWhitelistedModel } from "@app/lib/api/assistant/models";
 import config from "@app/lib/api/config";
 import { Authenticator } from "@app/lib/auth";
-import {
-  getAgentsDataRetention,
-  getConversationsDataRetention,
-} from "@app/lib/data_retention";
 import { DustError } from "@app/lib/error";
 import {
   ensureSlackNotificationsReady,
@@ -21,6 +17,13 @@ import {
   type NotificationAllowedTags,
 } from "@app/lib/notifications";
 import { renderEmail } from "@app/lib/notifications/email-templates/conversations-unread";
+import {
+  type ConversationDetailsPayload,
+  ConversationDetailsPayloadSchema,
+  ConversationDetailsSchema,
+  type ConversationDetailsType,
+  getConversationDetails,
+} from "@app/lib/notifications/helpers";
 import { ConversationResource } from "@app/lib/resources/conversation_resource";
 import { SpaceResource } from "@app/lib/resources/space_resource";
 import { UserMetadataModel } from "@app/lib/resources/storage/models/user";
@@ -29,16 +32,9 @@ import { concurrentExecutor } from "@app/lib/utils/async_utils";
 import { getConversationRoute } from "@app/lib/utils/router";
 import type { UserMessageOrigin } from "@app/types/assistant/conversation";
 import {
-  ConversationError,
   getConversationDisplayTitle,
-  isCompactionMessageType,
-  isLightAgentMessageType,
   isPodConversation,
-  isUserMessageType,
-  isVisibleMessage,
 } from "@app/types/assistant/conversation";
-import { isRichUserMention } from "@app/types/assistant/mentions";
-import { isContentFragmentType } from "@app/types/content_fragment";
 import type { NotificationCondition } from "@app/types/notification_preferences";
 import {
   CONVERSATION_NOTIFICATION_METADATA_KEYS,
@@ -54,10 +50,7 @@ import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import { assertNever } from "@app/types/shared/utils/assert_never";
 import { normalizeError } from "@app/types/shared/utils/error_utils";
-import {
-  decodeHtmlEntities,
-  stripMarkdown,
-} from "@app/types/shared/utils/markdown";
+import { stripMarkdown } from "@app/types/shared/utils/markdown";
 import { pluralize } from "@app/types/shared/utils/string_utils";
 import type { UserType } from "@app/types/user";
 import { workflow } from "@novu/framework";
@@ -65,18 +58,8 @@ import assert from "assert";
 import { Op } from "sequelize";
 import z from "zod";
 
-// When isNewProjectConversation is true, messageId is not required (the first
-// message is resolved from conversation content). Otherwise messageId is required.
-const ConversationUnreadPayloadSchema = z.object({
-  workspaceId: z.string(),
-  conversationId: z.string(),
-  messageId: z.string().optional(),
-  isNewProjectConversation: z.boolean().optional(),
-});
-
-export type ConversationUnreadPayloadType = z.infer<
-  typeof ConversationUnreadPayloadSchema
->;
+// The unread workflow operates on the shared conversation-details payload.
+export type ConversationUnreadPayloadType = ConversationDetailsPayload;
 
 export const shouldSendNotificationForAgentAnswer = (
   userMessageOrigin?: UserMessageOrigin | null
@@ -120,28 +103,6 @@ export const shouldSendNotificationForAgentAnswer = (
   }
 };
 
-const ConversationDetailsSchema = z.object({
-  subject: z.string(),
-  author: z.string(),
-  authorIsAgent: z.boolean(),
-  authorUserId: z.string().optional(),
-  isFromTrigger: z.boolean(),
-  isFromEmailAgentConversation: z.boolean(),
-  isFromSlackAgentConversation: z.boolean(),
-  workspaceName: z.string(),
-  mentionedUserIds: z.array(z.string()),
-  hasUnreadMessages: z.boolean(),
-  hasUnreadMentions: z.boolean(),
-  hasConversationRetentionPolicy: z.boolean(),
-  hasAgentRetentionPolicies: z.boolean(),
-  newMessageContent: z.string().nullable(),
-  // Fields for new project conversation notifications.
-  isNewProjectConversation: z.boolean().optional(),
-  projectName: z.string().optional(),
-});
-
-export type ConversationDetailsType = z.infer<typeof ConversationDetailsSchema>;
-
 // Wrapper for workflow step that may fail when conversation is deleted.
 const ConversationDetailsResultSchema = z.discriminatedUnion("success", [
   z.object({
@@ -156,202 +117,6 @@ const ConversationDetailsResultSchema = z.discriminatedUnion("success", [
 const UserNotificationDelaySchema = z.object({
   delay: z.enum(NOTIFICATION_DELAY_OPTIONS),
 });
-
-export const getConversationDetails = async ({
-  payload,
-  auth: providedAuth,
-  subscriberId,
-}: { payload: ConversationUnreadPayloadType } & (
-  | { auth: Authenticator; subscriberId?: never }
-  | { auth?: never; subscriberId: string }
-)): Promise<Result<ConversationDetailsType, ConversationError>> => {
-  if (!payload.isNewProjectConversation && !payload.messageId) {
-    throw new Error(
-      "messageId is required when isNewProjectConversation is false"
-    );
-  }
-
-  // Get or create auth from the discriminated union.
-  let auth: Authenticator;
-  if (providedAuth) {
-    auth = providedAuth;
-  } else {
-    // subscriberId may be empty when previewing the workflow step.
-    if (!subscriberId) {
-      return new Ok({
-        subject: "Deleted conversation",
-        author: "Deleted conversation",
-        authorIsAgent: false,
-        isFromTrigger: false,
-        isFromEmailAgentConversation: false,
-        isFromSlackAgentConversation: false,
-        workspaceName: "Deleted conversation",
-        mentionedUserIds: [],
-        avatarUrl: undefined,
-        hasUnreadMessages: false,
-        hasUnreadMentions: false,
-        hasConversationRetentionPolicy: false,
-        hasAgentRetentionPolicies: false,
-        newMessageContent: null,
-        isNewProjectConversation: false,
-      });
-    }
-    auth = await Authenticator.fromUserIdAndWorkspaceId(
-      subscriberId,
-      payload.workspaceId
-    );
-  }
-
-  const conversationRes = await getLightConversation(
-    auth,
-    payload.conversationId
-  );
-
-  if (conversationRes.isErr()) {
-    // Check if the conversation was deleted (expected during workflow delay).
-    const deletedConversation = await ConversationResource.fetchById(
-      auth,
-      payload.conversationId,
-      { includeDeleted: true }
-    );
-    if (deletedConversation) {
-      return new Err(new ConversationError("conversation_not_found"));
-    }
-    // Conversation never existed - unexpected.
-    throw new Error(`Conversation not found: ${payload.conversationId}`);
-  }
-
-  const conversation = conversationRes.value;
-
-  const workspaceName = auth.getNonNullableWorkspace().name;
-  // Decode HTML entities in conversation title (e.g. from email subjects
-  // that may contain &amp;, &lt;, etc.) so notification subjects/bodies
-  // display clean text.
-  const subject = decodeHtmlEntities(getConversationDisplayTitle(conversation));
-  const isFromTrigger = !!conversation.triggerId;
-
-  // Retrieve the message that triggered the notification.
-  // For new project conversations, use the first visible message
-  const message = !payload.isNewProjectConversation
-    ? conversation.content.find((msg) => msg.sId === payload.messageId)
-    : conversation.content.find(isVisibleMessage);
-  if (!message) {
-    // Message doesn't exist at all - could be true if it's in a branch.
-    return new Err(new ConversationError("message_not_found"));
-  }
-  if (message.visibility === "deleted") {
-    // Message was deleted during workflow delay - expected.
-    return new Err(new ConversationError("message_not_found"));
-  }
-
-  let author: string;
-  let authorIsAgent: boolean;
-  let authorUserId: string | undefined;
-  let mentionedUserIds: string[] = [];
-  const messageContent =
-    message.type === "agent_message" || message.type === "user_message"
-      ? message.content
-      : "";
-  const parentUserMessage = isLightAgentMessageType(message)
-    ? conversation.content
-        .flat()
-        .find((msg) => msg.sId === message.parentMessageId)
-    : undefined;
-  const isFromEmailAgentConversation =
-    (isUserMessageType(message) && message.context.origin === "email") ||
-    (parentUserMessage !== undefined &&
-      isUserMessageType(parentUserMessage) &&
-      parentUserMessage.context.origin === "email");
-
-  const isFromSlackAgentConversation =
-    (isUserMessageType(message) && message.context.origin === "slack") ||
-    (parentUserMessage !== undefined &&
-      isUserMessageType(parentUserMessage) &&
-      parentUserMessage.context.origin === "slack");
-
-  if (isCompactionMessageType(message)) {
-    // Compaction messages don't trigger notifications.
-    return new Err(new ConversationError("message_not_found"));
-  } else if (isContentFragmentType(message)) {
-    // Content fragments don't have author info.
-    author = "Someone else";
-    authorIsAgent = false;
-  } else if (isUserMessageType(message)) {
-    author =
-      message.user?.fullName ?? message.context.fullName ?? "Someone else";
-    authorUserId = message.user?.sId ?? undefined;
-    authorIsAgent = false;
-
-    // Extract approved user mentions from the rendered message.
-    mentionedUserIds = message.richMentions
-      .filter((m) => isRichUserMention(m) && m.status === "approved")
-      .map((m) => m.id);
-  } else if (isLightAgentMessageType(message)) {
-    author = message.configuration.name
-      ? `@${message.configuration.name}`
-      : "An agent";
-    authorIsAgent = true;
-  } else {
-    assertNever(message);
-  }
-
-  const unreadMessages = conversation.content.filter((msg) =>
-    isMessageUnread(msg, conversation.lastReadMs)
-  );
-
-  const hasUnreadMessages = unreadMessages.length > 0;
-
-  const hasUnreadMentions = unreadMessages.some((msg) => {
-    if (isContentFragmentType(msg) || isCompactionMessageType(msg)) {
-      return false;
-    }
-    return msg.richMentions.some(
-      (m) => isRichUserMention(m) && m.id === subscriberId
-    );
-  });
-
-  const conversationsRetention = await getConversationsDataRetention(auth);
-  const hasConversationRetentionPolicy = conversationsRetention !== null;
-
-  const agentsRetention = await getAgentsDataRetention(auth);
-  const hasAgentRetentionPolicies = conversation.content.some((msg) => {
-    if (msg.type !== "agent_message") {
-      return false;
-    }
-
-    return msg.configuration.sId in agentsRetention;
-  });
-
-  // Fetch project-specific details when this is a new project conversation notification.
-  let projectName: string | undefined;
-  const isNewProjectConversation = !!payload.isNewProjectConversation;
-
-  if (isNewProjectConversation && isPodConversation(conversation)) {
-    const project = await SpaceResource.fetchById(auth, conversation.spaceId);
-    if (project) {
-      projectName = project.name;
-    }
-  }
-
-  return new Ok({
-    subject,
-    author,
-    authorIsAgent,
-    authorUserId,
-    isFromTrigger,
-    isFromEmailAgentConversation,
-    isFromSlackAgentConversation,
-    workspaceName,
-    mentionedUserIds,
-    hasUnreadMessages,
-    hasUnreadMentions,
-    hasConversationRetentionPolicy,
-    hasAgentRetentionPolicies,
-    newMessageContent: messageContent,
-    isNewProjectConversation,
-    projectName,
-  });
-};
 
 const shouldSkipUnreadConversation = async ({
   subscriberId,
@@ -1081,7 +846,7 @@ export const conversationUnreadWorkflow = workflow(
     );
   },
   {
-    payloadSchema: ConversationUnreadPayloadSchema,
+    payloadSchema: ConversationDetailsPayloadSchema,
     tags: ["conversations"] as NotificationAllowedTags,
   }
 );
